Add option choose provider and restructure codebase based on providers.
Issue: https://github.com/sychhq/sych-llm-playground/issues/20
