+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "sych-llm-playground"
-version = "0.1.1"
+version = "0.2.0"
 description = "Sych LLM Playground is a command line tool deploy and interact with language models on the cloud."
 authors = ["Sych Inc. <dev@sych.io>"]
 license = "Apache-2.0"
