+++ src/sych_llm_playground/__main__.py
@@ -14,10 +14,13 @@ from .list import list
 
 
 @click.group(
-    help="""\
-sych_llm_playground is a command-line tool to manage and
-interact with language models in the cloud.
-"""
+    help=(
+        "sych-llm-playground is a command-line tool to manage and "
+        "interact with language models on the cloud.\n\n"
+        "Brought to you by Sych. Visit us at https://sych.io.\n\n"
+        "Detailed documenation at "
+        "https://sych-llm-playground.readthedocs.io"
+    )  # Replace with your actual shortened URL
 )
 @click.version_option()
 def main() -> None:
