+++ src/sych_llm_playground/cleanup.py
@@ -1,117 +1,18 @@
-"""Module for cleaning up deployed resources.
+"""Module for cleaning up deployed resources with chosen provider.
 
-This module provides functions to delete deployed resources such as models
-and endpoints from the cloud. It includes functions to
-select resources and initiate their cleanup.
-
-Functions:
-    choose_resource: Helps the user select a resource for deletion.
-    cleanup_resource: Deletes the selected resource.
-    cleanup: Main command-line interface function for cleaning up resources.
+This module provides the main function to delete deployed resources.
 """
 
-from typing import Any
-from typing import cast
-
-import boto3
 import click
-import inquirer
 
-from .utils.credentials import load_credentials
-from .utils.loader import start_loader
-from .utils.loader import stop_loader
-from .utils.resources import get_resources
+from .utils.provider_selection import select_provider_and_call_function
 
 
-def choose_resource(resource_type: str, sagemaker_client: Any) -> str:
-    """Choose a specific resource for cleanup.
-
-    This function lists the available resources of the specified type and prompts
-    the user to select one for deletion.
-
-    Args:
-        resource_type (str): The type of resource to delete (e.g., "Model", "Endpoint").
-        sagemaker_client (Any): The client to use to interact with AWS SageMaker.
-
-    Returns:
-        str: The name of the selected resource.
-    """
-    resources = get_resources(resource_type, sagemaker_client)
-    if len(resources) == 0:
-        exit(0)
-
-    questions = [
-        inquirer.List(
-            resource_type.lower(),
-            message=f"Select a {resource_type.lower()} to cleanup:",
-            choices=resources,
-        ),
-    ]
-
-    answers = inquirer.prompt(questions)
-    return cast(str, answers[resource_type.lower()])
-
-
-def cleanup_resource(resource_type: str) -> None:
-    """Delete a selected resource.
-
-    This function deletes a selected resource of the given type, either a model
-    or an endpoint. It manages the deletion process, including error handling
-    and success messaging.
-
-    Args:
-        resource_type (str): The type of resource to delete (e.g., "Model", "Endpoint").
-    """
-    sagemaker_client = boto3.client("sagemaker")
-    selected_resource = choose_resource(resource_type, sagemaker_client)
-
-    try:
-        loader_thread = start_loader(
-            message=f"Deleting selected {resource_type.lower()}...", color="green"
-        )
-        if resource_type == "Model":
-            sagemaker_client.delete_model(ModelName=selected_resource)
-        elif resource_type == "Endpoint":
-            sagemaker_client.delete_endpoint(EndpointName=selected_resource)
-        stop_loader(loader_thread)
-
-    except Exception as e:
-        stop_loader(loader_thread)
-        click.secho(
-            f"An error occurred while deleting {resource_type.lower()}s: {e}", fg="red"
-        )
-        exit(1)
-
-    click.secho(
-        f"{resource_type} {selected_resource} cleaned up successfully.", fg="green"
-    )
-
-
-@click.command(help="Remove deployed models and endpoints from the cloud.")
+@click.command(help="Remove deployed resources.")
 def cleanup() -> None:
-    """Command-line interface for cleaning up resources.
+    """Clean up resources with the chosen provider.
 
-    This function is the entry point for the CLI to remove deployed models
-    and endpoints. It loads credentials, asks the user what they'd like to clean up,
-    and then calls the appropriate cleanup function.
+    This function initiates the cleanup process for deployed resources
+    by calling the provider-specific function.
     """
-    load_credentials()
-
-    questions = [
-        inquirer.List(
-            "cleanup_options",
-            message="What would you like to cleanup?",
-            choices=[
-                "Model",
-                "Endpoint",
-            ],
-        ),
-    ]
-
-    answers = inquirer.prompt(questions)
-    selected_option = answers["cleanup_options"]
-
-    if selected_option == "Model":
-        cleanup_resource("Model")
-    elif selected_option == "Endpoint":
-        cleanup_resource("Endpoint")
+    select_provider_and_call_function("cleanup")
