+++ src/sych_llm_playground/providers/aws/__init__.py
@@ -1,0 +1,1 @@
+"""Sych LLM Playground Provider AWS."""
